+++ tests/dummy.source2/dummy/dummy2.py
@@ -1,0 +1,5 @@
+def baz():
+    c = 'c'
+
+def bat():
+    pass